add generic parser
