+++ src/ElectionResults.Core/Scheduler/CsvGenericDownloadJob.cs
@@ -1,0 +1,67 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.IO;
+using System.Linq;
+using System.Net;
+using System.Net.Http;
+using System.Text;
+using System.Threading.Tasks;
+using CSharpFunctionalExtensions;
+using CsvHelper;
+using CsvHelper.Configuration;
+using CsvHelper.Configuration.Attributes;
+using ElectionResults.Core.Endpoints.Query;
+using ElectionResults.Core.Endpoints.Response;
+using ElectionResults.Core.Entities;
+using ElectionResults.Core.Extensions;
+using ElectionResults.Core.Infrastructure;
+using ElectionResults.Core.Repositories;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+using Newtonsoft.Json;
+using Z.EntityFramework.Extensions;
+
+namespace ElectionResults.Core.Scheduler
+{
+    public abstract class CsvGenericDownloadJob : ICsvDownloaderJob
+    {
+        protected readonly IServiceProvider _serviceProvider;
+        protected HttpClient _httpClient;
+        protected string _csvUrl;
+
+        public CsvGenericDownloadJob(IServiceProvider serviceProvider)
+        {
+            _serviceProvider = serviceProvider;
+            _httpClient = new HttpClient();
+        }
+
+        public async Task DownloadFiles()
+        {
+            var stream = await DownloadFile(_csvUrl);
+            await ProcessStream(stream);
+        }
+
+        protected async Task<Stream> DownloadFile(string url)
+        {
+            try
+            {
+                var httpClientHandler = new HttpClientHandler
+                {
+                    AutomaticDecompression = DecompressionMethods.GZip | DecompressionMethods.Deflate
+                };
+                httpClientHandler.ServerCertificateCustomValidationCallback = (message, certificate2, arg3, arg4) => true;
+                _httpClient = new HttpClient(httpClientHandler);
+                var response = await _httpClient.GetStringAsync(url);
+                return new MemoryStream(Encoding.UTF8.GetBytes(response));
+            }
+            catch (Exception e)
+            {
+                Log.LogError(e, $"Failed to download file: {url}");
+                throw;
+            }
+        }
+
+        protected abstract Task ProcessStream(Stream stream);
+    }
+}
